+++ CHANGELOG.md
@@ -4,6 +4,34 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+## [0.3.0] - 04-09-2026
+
+### Added
+
+- Added `findLessonById()` to `LessonRepository`.
+- Added `getLessonsHistory()` to provide an array of lessons and history entries.
+- Pass `lessonRepository` to stats view handler.
+- Implemented `switchStatsView()`, `buildOverallView()`, `buildLessonsHistoryView()` inside stats view.
+- Added event listener for `overallStatsBtn` and `lessonsHistoryBtn` buttons.
+- Accepted `lessonRepository` as arguments inside stats view functions.
+- Added `about` and `credits` section in the settings view.
+- Added export to the `Difficulty` type.
+- Added event listeners for `lessonSearch`, and `difficultyFilter` elements.
+- Implemented `getFilteredLesson()` instead of the global lessons array to filter lessons by title or author.
+- Added `lessonSearch` and `difficultyFilter` elements in the HTML file, and added style.
+- Added helper text for empty search results.
+- Installed the `plugin-updater`, and `plugin-process`.
+- Added `checkForUpdates()` and `installUpdate()` functions and inject them inside the settings view.
+- Implemented event listener to track the installation flow
+
+### Changed
+
+- Divided the statistics page into two sections: **overall stats** and **lessons history stats**.
+- Changed the lessons source to the main entry URL.
+- Prevented typing listener when typing view is hidden.
+
+---
+
 ## [0.2.0] - 01-09-2026
 
 ### Added